Change to the homepage fetch the definitions
